move DOM hint tests to reportError.test.ts; tighten name-substitution assertion
